+++ app/src/main/java/woowacourse/movie/ui/home/MovieHomeKey.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.main
+package woowacourse.movie.ui.home
 
 object MovieHomeKey {
     const val ID = "movie_content_id"
