+++ app/src/main/res/layout/activity_movie_home.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MovieHomeActivity">
+    tools:context=".ui.home.MovieHomeActivity">
 
     <ListView
         android:id="@+id/movie_content_list"
